Moved files in db and server to their own src folders + experimented with outFile param
