--- db/lib/connect.d.ts
@@ -1,6 +1,0 @@
-/// <reference types="knex" />
-import * as Knex from 'knex';
-import { Model } from 'objection';
-declare const knex: Knex;
-export { knex };
-export { Model };
