--- db/lib/models/index.d.ts
@@ -1,1 +1,0 @@
-export { Note } from './note';
