--- db/lib/models/index.js
@@ -1,5 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-var note_1 = require("./note");
-exports.Note = note_1.Note;
-//# sourceMappingURL=index.js.map
