--- db/lib/models/note.d.ts
@@ -1,5 +1,0 @@
-import { BaseModel } from 'db-starter';
-export declare class Note extends BaseModel {
-    text: string;
-    static readonly tableName: string;
-}
