+++ db/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "db",
+  "name": "reperio-example-db",
   "version": "1.0.0",
   "description": "",
   "main": "./lib/index.js",
