+++ db/src/models/note.ts
